+++ cmd/repo-updater/repos/store_test.go
@@ -14,6 +14,7 @@ import (
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/sourcegraph/sourcegraph/cmd/repo-updater/repos"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/bitbucketserver"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
@@ -75,6 +76,14 @@ func testStoreListExternalServices(store repos.Store) func(*testing.T) {
 		UpdatedAt:   now,
 	}
 
+	awsCodeCommit := repos.ExternalService{
+		Kind:        "AWSCODECOMMIT",
+		DisplayName: "AWS CodeCommit - Test",
+		Config:      `{"region": "us-west-1"}`,
+		CreatedAt:   now,
+		UpdatedAt:   now,
+	}
+
 	otherService := repos.ExternalService{
 		Kind:        "OTHER",
 		DisplayName: "Other code hosts",
@@ -95,6 +104,7 @@ func testStoreListExternalServices(store repos.Store) func(*testing.T) {
 		&github,
 		&gitlab,
 		&bitbucketServer,
+		&awsCodeCommit,
 		&otherService,
 		&gitoliteService,
 	}
@@ -219,6 +229,14 @@ func testStoreUpsertExternalServices(store repos.Store) func(*testing.T) {
 			UpdatedAt:   now,
 		}
 
+		awsCodeCommit := repos.ExternalService{
+			Kind:        "AWSCODECOMMIT",
+			DisplayName: "AWS CodeCommit - Test",
+			Config:      `{"region": "us-west-1"}`,
+			CreatedAt:   now,
+			UpdatedAt:   now,
+		}
+
 		otherService := repos.ExternalService{
 			Kind:        "OTHER",
 			DisplayName: "Other code hosts",
@@ -239,6 +257,7 @@ func testStoreUpsertExternalServices(store repos.Store) func(*testing.T) {
 			&github,
 			&gitlab,
 			&bitbucketServer,
+			&awsCodeCommit,
 			&otherService,
 			&gitoliteService,
 		}
@@ -322,6 +341,7 @@ func testStoreUpsertRepos(store repos.Store) func(*testing.T) {
 			"github",
 			"gitlab",
 			"bitbucketserver",
+			"awscodecommit",
 			"other",
 			"gitolite",
 		}
@@ -386,6 +406,26 @@ func testStoreUpsertRepos(store repos.Store) func(*testing.T) {
 			Metadata: new(bitbucketserver.Repo),
 		}
 
+		awsCodeCommit := repos.Repo{
+			Name:        "git-codecommit.us-west-1.amazonaws.com/stripe-go",
+			Description: "The description",
+			Language:    "barlang",
+			Enabled:     true,
+			CreatedAt:   now,
+			ExternalRepo: api.ExternalRepoSpec{
+				ID:          "f001337a-3450-46fd-b7d2-650c0EXAMPLE",
+				ServiceType: "awscodecommit",
+				ServiceID:   "arn:aws:codecommit:us-west-1:999999999999:",
+			},
+			Sources: map[string]*repos.SourceInfo{
+				"extsvc:4": {
+					ID:       "extsvc:4",
+					CloneURL: "git@git-codecommit.us-west-1.amazonaws.com/v1/repos/stripe-go",
+				},
+			},
+			Metadata: new(awscodecommit.Repository),
+		}
+
 		otherRepo := repos.Repo{
 			Name: "git-host.com/org/foo",
 			ExternalRepo: api.ExternalRepoSpec{
@@ -394,8 +434,8 @@ func testStoreUpsertRepos(store repos.Store) func(*testing.T) {
 				ServiceType: "other",
 			},
 			Sources: map[string]*repos.SourceInfo{
-				"extsvc:4": {
-					ID:       "extsvc:3",
+				"extsvc:5": {
+					ID:       "extsvc:5",
 					CloneURL: "https://git-host.com/org/foo",
 				},
 			},
@@ -423,6 +463,7 @@ func testStoreUpsertRepos(store repos.Store) func(*testing.T) {
 			&github,
 			&gitlab,
 			&bitbucketServer,
+			&awsCodeCommit,
 			&otherRepo,
 			&gitoliteRepo,
 		}
@@ -554,6 +595,22 @@ func testStoreListRepos(store repos.Store) func(*testing.T) {
 		Metadata: new(bitbucketserver.Repo),
 	}
 
+	awsCodeCommit := repos.Repo{
+		Name: "git-codecommit.us-west-1.amazonaws.com/stripe-go",
+		ExternalRepo: api.ExternalRepoSpec{
+			ID:          "f001337a-3450-46fd-b7d2-650c0EXAMPLE",
+			ServiceType: "awscodecommit",
+			ServiceID:   "arn:aws:codecommit:us-west-1:999999999999:",
+		},
+		Sources: map[string]*repos.SourceInfo{
+			"extsvc:4": {
+				ID:       "extsvc:4",
+				CloneURL: "git@git-codecommit.us-west-1.amazonaws.com/v1/repos/stripe-go",
+			},
+		},
+		Metadata: new(awscodecommit.Repository),
+	}
+
 	otherRepo := repos.Repo{
 		Name: "git-host.com/org/foo",
 		ExternalRepo: api.ExternalRepoSpec{
@@ -591,6 +648,7 @@ func testStoreListRepos(store repos.Store) func(*testing.T) {
 		&github,
 		&gitlab,
 		&bitbucketServer,
+		&awsCodeCommit,
 		&otherRepo,
 		&gitoliteRepo,
 	}
@@ -599,6 +657,7 @@ func testStoreListRepos(store repos.Store) func(*testing.T) {
 		"github",
 		"gitlab",
 		"bitbucketserver",
+		"awscodecommit",
 		"other",
 		"gitolite",
 	}