+++ cmd/repo-updater/repos/types_test.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/sourcegraph/sourcegraph/pkg/api"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/bitbucketserver"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
@@ -57,6 +58,20 @@ func TestExternalService_Exclude(t *testing.T) {
 			"username: "admin",
 			"token": "secret",
 			"repositoryQuery": ["none"]
+		}`,
+		CreatedAt: now,
+		UpdatedAt: now,
+	}
+
+	awsCodeCommitService := ExternalService{
+		ID:          9,
+		Kind:        "AWSCODECOMMIT",
+		DisplayName: "AWS CodeCommit",
+		Config: `{
+			"region": "us-west-1",
+			"accessKeyID": "secret-accessKeyID",
+			"secretAccessKey": "secret-secretAccessKey",
+			"gitCredentials": {"username": "user", "password": "pw"},
 		}`,
 		CreatedAt: now,
 		UpdatedAt: now,
@@ -130,6 +145,18 @@ func TestExternalService_Exclude(t *testing.T) {
 			},
 		},
 		{
+			Metadata: &awscodecommit.Repository{
+				ID:   "f001337a-3450-46fd-b7d2-650c0EXAMPLE",
+				Name: "foo",
+			},
+		},
+		{
+			Metadata: &awscodecommit.Repository{
+				ID:   "b4455554-4444-5555-b7d2-888c9EXAMPLE",
+				Name: "baz",
+			},
+		},
+		{
 			Name: "git-host.mycorp.com/org/foo",
 			ExternalRepo: api.ExternalRepoSpec{
 				ID:          "1",
@@ -192,6 +219,20 @@ func TestExternalService_Exclude(t *testing.T) {
 					]
 				}`)
 			}),
+			awsCodeCommitService.With(func(e *ExternalService) {
+				e.Config = formatJSON(t, `
+				{
+					// Some comment
+					"region": "us-west-1",
+					"accessKeyID": "secret-accessKeyID",
+					"secretAccessKey": "secret-secretAccessKey",
+					"gitCredentials": {"username": "user", "password": "pw"},
+					"exclude": [
+						{"id": "f001337a-3450-46fd-b7d2-650c0EXAMPLE"},
+						{"name": "baz"}
+					]
+				}`)
+			}),
 			gitoliteService.With(func(e *ExternalService) {
 				e.Config = formatJSON(t, `
 				{
@@ -249,6 +290,19 @@ func TestExternalService_Exclude(t *testing.T) {
 					"repositoryQuery": ["none"],
 					"exclude": [
 						{"name": "org/boo"},
+					]
+				}`)
+			}),
+			awsCodeCommitService.With(func(e *ExternalService) {
+				e.Config = formatJSON(t, `
+				{
+					// Some comment
+					"region": "us-west-1",
+					"accessKeyID": "secret-accessKeyID",
+					"secretAccessKey": "secret-secretAccessKey",
+					"gitCredentials": {"username": "user", "password": "pw"},
+					"exclude": [
+						{"name": "boo"}
 					]
 				}`)
 			}),
@@ -321,6 +375,21 @@ func TestExternalService_Exclude(t *testing.T) {
 							{"name": "org/boo"},
 							{"id": 1, "name": "org/foo"},
 							{"name": "org/baz"}
+						]
+					}`)
+				}),
+				awsCodeCommitService.With(func(e *ExternalService) {
+					e.Config = formatJSON(t, `
+					{
+						// Some comment
+						"region": "us-west-1",
+						"accessKeyID": "secret-accessKeyID",
+						"secretAccessKey": "secret-secretAccessKey",
+						"gitCredentials": {"username": "user", "password": "pw"},
+						"exclude": [
+							{"name": "boo"},
+							{"id": "f001337a-3450-46fd-b7d2-650c0EXAMPLE", "name": "foo"},
+							{"id": "b4455554-4444-5555-b7d2-888c9EXAMPLE", "name": "baz"}
 						]
 					}`)
 				}),